(PC-5884) Move SignUp from context to the api file
