+++ src/features/auth/__mocks__/api.ts
@@ -1,3 +1,7 @@
 export const signIn = jest.fn()
 
 export const useSignIn = jest.fn(() => signIn)
+
+export const signUp = jest.fn()
+
+export const useSignUp = jest.fn(() => signUp)
