+++ src/features/auth/login/Login.test.tsx
@@ -27,7 +27,7 @@ const mockUseSignIn = useSignIn as jest.Mock
 function renderLogin() {
   return render(
     <AuthContext.Provider
-      value={{ isLoggedIn: true, setIsLoggedIn: jest.fn(), signUp: jest.fn(), signOut: jest.fn() }}>
+      value={{ isLoggedIn: true, setIsLoggedIn: jest.fn(), signOut: jest.fn() }}>
       <Login />
     </AuthContext.Provider>
   )
